requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
SQLAlchemy>=1.4
alembic>=1.12
psycopg2-binary>=2.9
passlib[argon2]>=1.7
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3
pydantic>=2.0
pydantic-settings>=2.0
email-validator>=2.0
python-dotenv>=1.0
slowapi>=0.1.8
python-multipart>=0.0.7
